fewer warnings compiling for Win64

+++ include/md4.h
@@ -1,7 +1,7 @@
 #ifndef MD4_H
 #define MD4_H
 
-/* $Id: md4.h,v 1.7 2007-03-12 13:28:50 freddy77 Exp $ */
+/* $Id: md4.h,v 1.8 2009-01-16 20:27:56 jklowden Exp $ */
 
 #if defined(__GNUC__) && __GNUC__ >= 4
 #pragma GCC visibility push(hidden)
@@ -15,7 +15,7 @@ struct MD4Context
 };
 
 void MD4Init(struct MD4Context *context);
-void MD4Update(struct MD4Context *context, unsigned char const *buf, unsigned len);
+void MD4Update(struct MD4Context *context, unsigned char const *buf, size_t len);
 void MD4Final(struct MD4Context *context, unsigned char *digest);
 void MD4Transform(TDS_UINT buf[4], TDS_UINT const in[16]);
